Attach the command line interface to the interpreter
And use a single threaded polling approach for the CLI. I had the CLI
implemented in a separate thread, however, the program would seg fault
if a graphics call was made. I could rework to do all interpreter work
on the main thread. That would need more elaborate thread
synchronization as I would want the CLI to wait for a response before
showing the prompt for the next line of input. I'll go ahead with a
single thread for now and rethink in the future if there are problems.